requirements.txt
flask
flask-cors
flask-mysqldb
requests
google-auth
google-auth-oauthlib
google-api-python-client
mysql-connector-python
cryptography
flet
